# perf(node/swc): switch to mimalloc