--- a/babel/compat/benches/babelify.rs
+++ b/babel/compat/benches/babelify.rs
@@ -1,4 +1,5 @@
 #![feature(test)]
+#![feature(bench_black_box)]
 
 extern crate test;