--- a/benches/typescript.rs
+++ b/benches/typescript.rs
@@ -1,4 +1,5 @@
 #![feature(test)]
+#![feature(bench_black_box)]
 
 extern crate test;