--- a/bundler/tests/fixture.rs
+++ b/bundler/tests/fixture.rs
@@ -44,6 +44,8 @@ fn add_test<F: FnOnce() + Send + 'static>(
             ignore,
             should_panic: No,
             allow_fail: false,
+            compile_fail: false,
+            no_run: false,
         },
         testfn: DynTestFn(Box::new(f)),
     });