--- a/ecmascript/codegen/benches/with_parse.rs
+++ b/ecmascript/codegen/benches/with_parse.rs
@@ -1,4 +1,5 @@
 #![feature(test)]
+#![feature(bench_black_box)]
 
 extern crate test;