--- a/ecmascript/codegen/src/macros.rs
+++ b/ecmascript/codegen/src/macros.rs
@@ -26,10 +26,10 @@ macro_rules! emit {
 
 macro_rules! keyword {
     ($emitter:expr, $span:expr, $s:expr) => {
-        $emitter.wr.write_keyword(Some($span), $s)?;
+        $emitter.wr.write_keyword(Some($span), $s)?
     };
     ($emitter:expr, $s:expr) => {
-        $emitter.wr.write_keyword(None, $s)?;
+        $emitter.wr.write_keyword(None, $s)?
     };
 }
 
@@ -42,10 +42,10 @@ macro_rules! punct {
     };
 
     ($emitter:expr, ";") => {
-        $emitter.wr.write_semi(None)?;
+        $emitter.wr.write_semi(None)?
     };
     ($emitter:expr, $s:expr) => {
-        $emitter.wr.write_punct(None, $s)?;
+        $emitter.wr.write_punct(None, $s)?
     };
 }
 
@@ -61,7 +61,7 @@ macro_rules! operator {
 
 macro_rules! space {
     ($emitter:expr) => {
-        $emitter.wr.write_space()?;
+        $emitter.wr.write_space()?
     };
     ($emitter:expr,) => {
         space!($emitter)