--- a/ecmascript/parser/benches/lexer.rs
+++ b/ecmascript/parser/benches/lexer.rs
@@ -1,4 +1,5 @@
 #![feature(test)]
+#![feature(bench_black_box)]
 
 extern crate test;