--- a/ecmascript/parser/tests/jsx.rs
+++ b/ecmascript/parser/tests/jsx.rs
@@ -33,6 +33,8 @@ fn add_test<F: FnOnce() + Send + 'static>(
             ignore,
             should_panic: No,
             allow_fail: false,
+            compile_fail: false,
+            no_run: false,
         },
         testfn: DynTestFn(Box::new(f)),
     });