--- a/ecmascript/parser/tests/span.rs
+++ b/ecmascript/parser/tests/span.rs
@@ -29,6 +29,8 @@ fn add_test<F: FnOnce() + Send + 'static>(
             ignore,
             should_panic: No,
             allow_fail: false,
+            compile_fail: false,
+            no_run: false,
         },
         testfn: DynTestFn(Box::new(f)),
     });