--- a/ecmascript/preset-env/tests/test.rs
+++ b/ecmascript/preset-env/tests/test.rs
@@ -160,6 +160,8 @@ fn load() -> Result<Vec<TestDescAndFn>, Error> {
                 name: TestName::DynTestName(name),
                 allow_fail: false,
                 should_panic: ShouldPanic::No,
+                compile_fail: false,
+                no_run: false,
             },
             testfn: TestFn::DynTestFn(Box::new(move || {
                 //