--- a/ecmascript/transforms/base/tests/fixer_test262.rs
+++ b/ecmascript/transforms/base/tests/fixer_test262.rs
@@ -104,6 +104,8 @@ fn add_test<F: FnOnce() + Send + 'static>(
             ignore,
             should_panic: No,
             allow_fail: false,
+            compile_fail: false,
+            no_run: false,
         },
         testfn: DynTestFn(Box::new(f)),
     });