--- a/node/base/src/lib.rs
+++ b/node/base/src/lib.rs
@@ -3,15 +3,9 @@
 //! The swc crates related to the  node binding should depend on this crate.
 
 #[cfg(all(
-    unix,
+    target_arch = "x86_64",
     not(target_env = "musl"),
-    not(target_os = "freebsd"),
-    not(target_arch = "arm"),
-    not(target_arch = "aarch64")
+    not(debug_assertions)
 ))]
 #[global_allocator]
-static ALLOC: jemallocator::Jemalloc = jemallocator::Jemalloc;
-
-#[cfg(all(windows, not(target_arch = "aarch64")))]
-#[global_allocator]
-static ALLOC: mimalloc::MiMalloc = mimalloc::MiMalloc;
+static ALLOC: mimalloc_rust::GlobalMiMalloc = mimalloc_rust::GlobalMiMalloc;