--- a/spack/benches/bench.rs
+++ b/spack/benches/bench.rs
@@ -1,6 +1,7 @@
 //! This benchmark lives here because I don't want to implement `Resolve` again
 //! and again.
 #![feature(test)]
+#![feature(bench_black_box)]
 
 extern crate test;