--- a/spack/tests/fixture.rs
+++ b/spack/tests/fixture.rs
@@ -42,6 +42,8 @@ fn add_test<F: FnOnce() + Send + 'static>(
             ignore,
             should_panic: No,
             allow_fail: false,
+            compile_fail: false,
+            no_run: false,
         },
         testfn: DynTestFn(Box::new(f)),
     });