### v1.2.76

```
@swc-node/core x 396 ops/sec ±1.83% (90 runs sampled)
esbuild x 465 ops/sec ±1.00% (88 runs sampled)
typescript x 26.83 ops/sec ±10.61% (52 runs sampled)
babel x 29.41 ops/sec ±9.19% (55 runs sampled)
Transform rxjs/AjaxObservable.ts benchmark bench suite: Fastest is esbuild
@swc-node/core x 1,023 ops/sec ±1.84% (72 runs sampled)
esbuild x 948 ops/sec ±1.17% (80 runs sampled)
Transform rxjs/AjaxObservable.ts parallel benchmark bench suite: Fastest is @swc-node/core
```

### This commit

```
@swc-node/core x 436 ops/sec ±0.57% (91 runs sampled)
esbuild x 468 ops/sec ±1.67% (87 runs sampled)
typescript x 27.01 ops/sec ±11.05% (51 runs sampled)
babel x 29.40 ops/sec ±8.73% (55 runs sampled)
Transform rxjs/AjaxObservable.ts benchmark bench suite: Fastest is esbuild
@swc-node/core x 1,089 ops/sec ±1.25% (75 runs sampled)
esbuild x 967 ops/sec ±1.14% (82 runs sampled)
Transform rxjs/AjaxObservable.ts parallel benchmark bench suite: Fastest is @swc-node/core
```