@kdy1 2021-08-13
Wow, it's a big difference.
Is mimalloc always a win for any target?