@Brooooooklyn 2021-08-13
It should be compatible for all x86  platforms and macOS-m1.