@Brooooooklyn 2021-08-13
`cargo +$(cat rust-toolchain) doc` should work

```diff
    uses: actions-rs/toolchain@v1
        with:
          profile: minimal
--        toolchain: nightly-2021-03-30
```

This line could be removed too, because `actions-rs/toolchain` supported the `rust-toolchain` file.