@kdy1 2021-08-13
Yeah, you are correct. I'll remove the line.
Btw, is there any reason you used `mimalloc-rust ` instead of `mimalloc`?