@Brooooooklyn 2021-08-13
> The `mimalloc` had to dynamically check if the memory type was align every time it was allocated, but now I've taken a closer look at Rust's ABI and found that all Rust types seem to be Align, so I'll just use Align.
> The second reason is that the previous library may cause memory errors because there are many data types that are not covered, especially when the extend function is used.

/cc @LemonHX